refactor/ovos-core!!
this makes setup.py provide 2 packages when installing core

minimal needed code to run the skills service lives in ovos_core

goodbye `from mycroft import *`!

the mycroft module remains only for backwards compatibility, downstream should start importing `ovos_core`
