+++ .github/workflows/unit_tests.yml
@@ -50,7 +50,7 @@ jobs:
           python -m pip install build wheel
       - name: Install core repo
         run: |
-          pip install .[audio,mark1,stt,tts,skills,gui,bus,PHAL,all,deprecated]
+          pip install .[mycroft,lgpl,deprecated]
       - name: Install test dependencies
         run: |
           pip install -r requirements/tests.txt
