+++ mycroft/__init__.py
@@ -32,3 +32,7 @@ _log_level = _cfg.get("log_level", "INFO")
 _logs_conf = _cfg.get("logs") or {}
 _logs_conf["level"] = _log_level
 LOG.init(_logs_conf)  # read log level from config
+
+
+LOG.warning("mycroft has been deprecated! please start importing from ovos_core and companion packages\n"
+            "mycroft module remains available for backwards compatibility and will be removed in version 0.2.0")
