+++ mycroft/skills/audioservice.py
@@ -1,2 +1,2 @@
 # this was moved in order to allow importing it without dragging skills service dependencies
-from mycroft.audio.interface import AudioService
+from ovos_utils.skills.audioservice import ClassicAudioServiceInterface as AudioService
