+++ mycroft/skills/common_play_skill.py
@@ -17,7 +17,7 @@ from enum import Enum, IntEnum
 from abc import ABC, abstractmethod
 from ovos_bus_client.message import Message
 from ovos_workshop.skills.mycroft_skill import MycroftSkill
-from mycroft.skills.audioservice import AudioService
+from ovos_utils.skills.audioservice import ClassicAudioServiceInterface as AudioService
 
 
 class CPSMatchLevel(Enum):
