--- requirements/extra-PHAL.txt
@@ -1,1 +1,0 @@
-ovos-PHAL~=0.0, >=0.0.5a2
