--- requirements/extra-audiobackend.txt
@@ -1,2 +1,0 @@
-# audio service
-ovos-audio
