--- requirements/extra-deprecated.txt
@@ -1,9 +1,0 @@
-# only required if importing from mycroft.deprecated module
-# not used internally
-msm
-mock_msm>=0.9
-ovos_cli_client
-python-vlc>=1.1.2
-pyalsaaudio>=0.8
-pyserial>=3.0
-pillow>=8.3
