--- requirements/extra-gui.txt
@@ -1,2 +1,0 @@
-# gui service
-ovos-gui~=0.0, >=0.0.2
