--- requirements/extra-skills-lgpl.txt
@@ -1,7 +1,0 @@
-adapt-parser~=0.5
-padacioso~=0.1.2
-ovos-lingua-franca~=0.4, >=0.4.2
-PyYAML~=5.4
-ovos_workshop<0.1.0, >=0.0.12a20
-padatious~=0.4.8
-fann2==1.0.7
