+++ requirements/lgpl.txt
@@ -1,0 +1,3 @@
+padatious>=0.4.8, < 0.5.0
+fann2>=1.0.7, < 1.1.0
+padaos>=0.1, < 0.2
