--- requirements/minimal.txt
@@ -1,15 +1,0 @@
-requests~=2.26
-python-dateutil~=2.6
-watchdog~=2.1
-
-combo-lock~=0.2
-padacioso~=0.1.2
-
-
-ovos_bus_client<0.1.0, >=0.0.4a3
-ovos-utils<0.1.0, >=0.0.33a3
-ovos-plugin-manager<0.1.0, >=0.0.23a9
-ovos-config~=0.0,>=0.0.8
-ovos-lingua-franca>=0.4.7
-ovos_backend_client<0.1.0, >=0.0.6
-ovos_workshop<0.1.0, >=0.0.12a20
