+++ requirements/requirements.txt
@@ -1,37 +1,15 @@
-requests~=2.26
-tornado~=6.0, >=6.0.3
-psutil~=5.6.6
-python-dateutil~=2.6
+requests>=2.26, <3.0
+python-dateutil>=2.6, <3.0
+watchdog>=2.1, <3.0
 combo-lock>=0.2.2, <0.3
-PyYAML~=5.4
-watchdog~=2.1
 
-adapt-parser~=0.5
-padatious~=0.4.8
-padacioso~=0.1.2
-fann2==1.0.7
-padaos~=0.1
+padacioso>=0.1.2, <0.2.0
+adapt-parser>=1.0.0, <2.0.0
 
 ovos_bus_client<0.1.0, >=0.0.4a3
-ovos_backend_client<0.1.0, >=0.0.6
-ovos-config~=0.0,>=0.0.8
-ovos-utils<0.1.0, >=0.0.33a3
+ovos-utils<0.1.0, >=0.0.33a5
 ovos-plugin-manager<0.1.0, >=0.0.23a9
-ovos_workshop<0.1.0, >=0.0.12a20
-ovos_PHAL<0.1.0, >=0.0.2
+ovos-config~=0.0,>=0.0.8
 ovos-lingua-franca>=0.4.7
-
-ovos-audio~=0.0, >=0.0.2a10
-ovos-listener~=0.0, >=0.0.2a9
-ovos-gui~=0.0, >=0.0.2
-ovos-messagebus~=0.0, >=0.0.2
-
-ovos-skill-volume>=0.0.1,<0.1.0
-ovos-skill-fallback-unknown>=0.0.2,<0.1.0
-ovos-skill-stop>=0.2,<0.3.0
-# neon-skill-alerts~=1.2
-ovos-skill-personal>=0.0.3,<0.1.0
-ovos-skill-naptime>=0.2.2,<0.3.0
-ovos-skill-date-time>=0.2.2,<0.3.0
-ovos-skill-hello-world>=0.0.3,<0.1.0
-ovos-skill-filebrowser>=0.0.2,<0.1.0
+ovos_backend_client<0.1.0, >=0.0.6
+ovos_workshop<0.1.0, >=0.0.12a17
