+++ scripts/bump_alpha.py
@@ -2,7 +2,7 @@ import fileinput
 from os.path import join, dirname
 
 
-version_file = join(dirname(dirname(__file__)), "mycroft", "version.py")
+version_file = join(dirname(dirname(__file__)), "ovos_core", "version.py")
 version_var_name = "OVOS_VERSION_ALPHA"
 
 with open(version_file, "r", encoding="utf-8") as v:
