+++ test/unittests/skills/test_mycroft_skill.py
@@ -25,7 +25,7 @@ from adapt.intent import IntentBuilder
 from ovos_config import Configuration
 from ovos_bus_client.message import Message
 from mycroft.skills.core import MycroftSkill, resting_screen_handler, intent_handler
-from mycroft.skills.intent_service import open_intent_envelope
+from ovos_utils.intents.intent_service_interface import open_intent_envelope
 from mycroft.skills.skill_data import (load_regex_from_file, load_regex,
                                        load_vocabulary, read_vocab_file)
 from test.util import base_config
